ho finito la gestione dell'area personale e iniziato la gestione dell'ordine

+++ src/main/java/controller/utente/LogoutServlet.java
@@ -7,6 +7,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import jakarta.servlet.http.HttpSession;
 import model.carrelloService.Carrello;
+import model.carrelloService.CarrelloDAO;
 import model.carrelloService.RigaCarrello;
 import model.carrelloService.RigaCarrelloDAO;
 import model.libroService.Libro;
@@ -19,20 +20,30 @@ import java.io.IOException;
 public class LogoutServlet extends HttpServlet {
     public void doGet(HttpServletRequest request, HttpServletResponse response) throws ServletException, IOException {
         HttpSession session = request.getSession();
-        Carrello carrello = (Carrello) session.getAttribute("Carrello");
-        WishList wishList = (WishList) session.getAttribute("WishList");
-        Utente utente = (Utente) session.getAttribute("Utente");
+        CarrelloDAO carrelloDAO = new CarrelloDAO();
+        WishListDAO wishListDAO = new WishListDAO();
+        Carrello carrello = (Carrello) session.getAttribute("carrello");
+        WishList wishList = (WishList) session.getAttribute("wishList");
+        Utente utente = (Utente) session.getAttribute("utente");
 
         try{
             RigaCarrelloDAO rigaCarrelloService = new RigaCarrelloDAO();
-            rigaCarrelloService.deleteRigheCarrelloByIdCarrello(carrello.getIdCarrello());//elimino ciò che è presente nel db
+            if(carrelloDAO.doRetriveByUtente(utente.getEmail()) != null && !(carrelloDAO.doRetriveByUtente(utente.getEmail()).getRigheCarrello().isEmpty())) {
+                rigaCarrelloService.deleteRigheCarrelloByIdCarrello(carrello.getIdCarrello());//elimino ciò che è presente nel db
+            }
             WishListDAO wishListService = new WishListDAO();
-            wishListService.deleteWishListByEmail(utente.getEmail());//elimino ciò che è presente nel db
+            if(wishListDAO.doRetrieveByEmail(utente.getEmail())!= null && !(wishListDAO.doRetrieveByEmail(utente.getEmail()).getLibri().isEmpty())) {
+                wishListService.deleteWishListByEmail(utente.getEmail());//elimino ciò che è presente nel db
+            }
 
-            for(RigaCarrello riga : carrello.getRigheCarrello())
-                rigaCarrelloService.doSave(riga); //ripopolo il db con le informazioni presenti in sessione
-            for(Libro libro : wishList.getLibri()){
-                wishListService.doSave(wishList, libro.getIsbn()); //ripopolo il db con le informazioni presenti in sessione
+            if(carrello.getRigheCarrello()!= null) {
+                for (RigaCarrello riga : carrello.getRigheCarrello())
+                    rigaCarrelloService.doSave(riga); //ripopolo il db con le informazioni presenti in sessione
+            }
+            if(wishList.getLibri()!= null) {
+                for (Libro libro : wishList.getLibri()) {
+                    wishListService.doSave(wishList, libro.getIsbn()); //ripopolo il db con le informazioni presenti in sessione
+                }
             }
